Python/requirements.txt
blinker==1.9.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.8
Flask==3.1.0
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
MarkupSafe==3.0.2
numpy==1.26.4
pandas==2.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
schedule==1.2.2
six==1.16.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.1.3
